added proper packages for dto request and responses and also refactored correcly.
