+++ src/main/java/com/AirBnb/TimberAndStone/controllers/AuthController.java
@@ -1,9 +1,9 @@
 package com.AirBnb.TimberAndStone.controllers;
 
-import com.AirBnb.TimberAndStone.dto.AuthRequest;
-import com.AirBnb.TimberAndStone.dto.AuthResponse;
-import com.AirBnb.TimberAndStone.dto.RegisterRequest;
-import com.AirBnb.TimberAndStone.dto.RegisterResponse;
+import com.AirBnb.TimberAndStone.requests.authentication.AuthRequest;
+import com.AirBnb.TimberAndStone.responses.authentication.AuthResponse;
+import com.AirBnb.TimberAndStone.requests.authentication.RegisterRequest;
+import com.AirBnb.TimberAndStone.responses.authentication.RegisterResponse;
 import com.AirBnb.TimberAndStone.services.UserService;
 import com.AirBnb.TimberAndStone.utils.JwtUtil;
 import jakarta.servlet.http.HttpServletResponse;
