+++ src/main/java/com/AirBnb/TimberAndStone/requests/authentication/AuthRequest.java
@@ -1,4 +1,4 @@
-package com.AirBnb.TimberAndStone.dto;
+package com.AirBnb.TimberAndStone.requests.authentication;
 
 import jakarta.validation.constraints.NotBlank;
