+++ src/main/java/com/AirBnb/TimberAndStone/responses/booking/AllBookingsResponse.java
@@ -1,4 +1,4 @@
-package com.AirBnb.TimberAndStone.dto;
+package com.AirBnb.TimberAndStone.responses.booking;
 
 import com.AirBnb.TimberAndStone.models.BookingStatus;
 import com.AirBnb.TimberAndStone.models.Period;
