+++ src/main/java/com/AirBnb/TimberAndStone/responses/booking/BookingResponse.java
@@ -1,42 +1,43 @@
-package com.AirBnb.TimberAndStone.dto;
+package com.AirBnb.TimberAndStone.responses.booking;
 
 import com.AirBnb.TimberAndStone.models.BookingStatus;
 import com.AirBnb.TimberAndStone.models.Period;
 
-public class PatchBookingResponse {
-    private String message;
-    private String rental;
+import java.time.LocalDateTime;
+
+public class BookingResponse {
     private String bookingNumber;
+    private String rental;
     private String guest;
     private Integer numberOfGuests;
     private Period period;
     private Double totalPrice;
-    private Boolean isPaid;
     private BookingStatus status;
     private String note;
+    private LocalDateTime bookingDate;
 
-    public PatchBookingResponse() {
+    public BookingResponse() {
     }
 
-    public PatchBookingResponse(String message, String rental, String bookingNumber, String guest, Integer numberOfGuests, Period period, Double totalPrice, Boolean isPaid, BookingStatus status, String note) {
-        this.message = message;
-        this.rental = rental;
+    public BookingResponse(String bookingNumber, String rental, String guest, Integer numberOfGuests, Period period, Double totalPrice, BookingStatus status, String note, LocalDateTime bookingDate) {
         this.bookingNumber = bookingNumber;
+        this.rental = rental;
         this.guest = guest;
         this.numberOfGuests = numberOfGuests;
         this.period = period;
         this.totalPrice = totalPrice;
-        this.isPaid = isPaid;
         this.status = status;
         this.note = note;
+        this.bookingDate = bookingDate;
     }
 
-    public String getMessage() {
-        return message;
+
+    public String getBookingNumber() {
+        return bookingNumber;
     }
 
-    public void setMessage(String message) {
-        this.message = message;
+    public void setBookingNumber(String bookingNumber) {
+        this.bookingNumber = bookingNumber;
     }
 
     public String getRental() {
@@ -45,14 +46,6 @@ public class PatchBookingResponse {
 
     public void setRental(String rental) {
         this.rental = rental;
-    }
-
-    public String getBookingNumber() {
-        return bookingNumber;
-    }
-
-    public void setBookingNumber(String bookingNumber) {
-        this.bookingNumber = bookingNumber;
     }
 
     public String getGuest() {
@@ -87,14 +80,6 @@ public class PatchBookingResponse {
         this.totalPrice = totalPrice;
     }
 
-    public Boolean getPaid() {
-        return isPaid;
-    }
-
-    public void setPaid(Boolean paid) {
-        isPaid = paid;
-    }
-
     public BookingStatus getStatus() {
         return status;
     }
@@ -109,5 +94,13 @@ public class PatchBookingResponse {
 
     public void setNote(String note) {
         this.note = note;
+    }
+
+    public LocalDateTime getBookingDate() {
+        return bookingDate;
+    }
+
+    public void setBookingDate(LocalDateTime bookingDate) {
+        this.bookingDate = bookingDate;
     }
 }
