+++ src/main/java/com/AirBnb/TimberAndStone/responses/rental/RentalFindByCategoryResponse.java
@@ -1,4 +1,4 @@
-package com.AirBnb.TimberAndStone.dto;
+package com.AirBnb.TimberAndStone.responses.rental;
 
 import com.AirBnb.TimberAndStone.models.Category;
