+++ src/main/java/com/AirBnb/TimberAndStone/responses/rental/RentalResponse.java
@@ -1,4 +1,4 @@
-package com.AirBnb.TimberAndStone.dto;
+package com.AirBnb.TimberAndStone.responses.rental;
 
 public class RentalResponse {
