+++ src/main/java/com/AirBnb/TimberAndStone/responses/rentalReview/RentalReviewResponse.java
@@ -1,23 +1,22 @@
-package com.AirBnb.TimberAndStone.dto;
+package com.AirBnb.TimberAndStone.responses.rentalReview;
 
-public class UserReviewResponse {
+public class RentalReviewResponse {
     private String message;
     private String user;
-    private String host;
     private String rental;
     private Integer rating;
     private String review;
 
-    public UserReviewResponse(String message, String user, String host, String rental, Integer rating, String review) {
+
+    public RentalReviewResponse() {
+    }
+
+    public RentalReviewResponse(String message, String user, String rental, Integer rating, String review) {
         this.message = message;
         this.user = user;
-        this.host = host;
         this.rental = rental;
         this.rating = rating;
         this.review = review;
-    }
-
-    public UserReviewResponse() {
     }
 
     public String getMessage() {
@@ -34,14 +33,6 @@ public class UserReviewResponse {
 
     public void setUser(String user) {
         this.user = user;
-    }
-
-    public String getHost() {
-        return host;
-    }
-
-    public void setHost(String host) {
-        this.host = host;
     }
 
     public String getRental() {
