+++ src/main/java/com/AirBnb/TimberAndStone/responses/userReview/UserReviewResponse.java
@@ -1,22 +1,23 @@
-package com.AirBnb.TimberAndStone.dto;
+package com.AirBnb.TimberAndStone.responses.userReview;
 
-public class RentalReviewResponse {
+public class UserReviewResponse {
     private String message;
     private String user;
+    private String host;
     private String rental;
     private Integer rating;
     private String review;
 
-
-    public RentalReviewResponse() {
-    }
-
-    public RentalReviewResponse(String message, String user, String rental, Integer rating, String review) {
+    public UserReviewResponse(String message, String user, String host, String rental, Integer rating, String review) {
         this.message = message;
         this.user = user;
+        this.host = host;
         this.rental = rental;
         this.rating = rating;
         this.review = review;
+    }
+
+    public UserReviewResponse() {
     }
 
     public String getMessage() {
@@ -33,6 +34,14 @@ public class RentalReviewResponse {
 
     public void setUser(String user) {
         this.user = user;
+    }
+
+    public String getHost() {
+        return host;
+    }
+
+    public void setHost(String host) {
+        this.host = host;
     }
 
     public String getRental() {
