+++ src/main/java/com/AirBnb/TimberAndStone/services/RentalReviewService.java
@@ -1,6 +1,6 @@
 package com.AirBnb.TimberAndStone.services;
 
-import com.AirBnb.TimberAndStone.dto.RentalReviewResponse;
+import com.AirBnb.TimberAndStone.responses.rentalReview.RentalReviewResponse;
 import com.AirBnb.TimberAndStone.exceptions.ConflictException;
 import com.AirBnb.TimberAndStone.exceptions.ResourceNotFoundException;
 import com.AirBnb.TimberAndStone.exceptions.UnauthorizedException;
