+++ src/main/java/com/AirBnb/TimberAndStone/services/UserReviewService.java
@@ -1,6 +1,6 @@
 package com.AirBnb.TimberAndStone.services;
 
-import com.AirBnb.TimberAndStone.dto.UserReviewResponse;
+import com.AirBnb.TimberAndStone.responses.userReview.UserReviewResponse;
 import com.AirBnb.TimberAndStone.exceptions.ConflictException;
 import com.AirBnb.TimberAndStone.exceptions.ResourceNotFoundException;
 import com.AirBnb.TimberAndStone.exceptions.UnauthorizedException;
